Add simple assertions
